Class update
Apologies for the multiple pull requests - wasn't sure whether to commit what I had previously done on the gradient-loop-etc branch to the master branch. Would you be able to have a look at what I've done for the curve fitting method? Pretty much just the basic idea - still need to have  abit of a play around but it works as it is with the example I've also uploaded!